refactor(audit): split AuditDrawer to clear no-giant-component + collapse dup export
AuditDrawer's component body was 536 lines, over react-doctor's hard 300-line
no-giant-component threshold. Extract cohesive module-scope pieces so the render
is byte-for-byte unchanged while each concern stands on its own:

- New audit-item-row.tsx: AuditItemRow (scanned row), AuditPendingRow (expected
  row), renderAuditItems (list composer), shared assetTypeBadgeClass.
- New siblings in audit-drawer.tsx: AuditDrawerTitle, AuditDrawerEmptyState.
- Collapse the duplicate `export function AuditDrawer` + `export default` into a
  single default export (clears the knip duplicate/unused-export warnings).

Pure structural refactor, reviewed for behavior equivalence (prop threading,
classNames, keys, AnimatePresence fragment contract, and title string-vs-element
semantics all preserved). react-doctor exit 0 (warning gone); typecheck clean.
